fixed sudden problem rating drops, fixed problems reset to unsolved and set resets

+++ View/TagNames/view.ctp
@@ -9,8 +9,11 @@
 	<?php } ?>
 	<p>Created by <?php echo $tn['TagName']['user'] ?>.</p>
 	<?php if($_SESSION['loggedInUser']['User']['isAdmin']>0){ ?>
-		<a href="/tag_names/edit/<?php echo $tn['TagName']['id']; ?>">Edit</a> | 
-		<a href="/tag_names/delete/<?php echo $tn['TagName']['id']; ?>">Delete</a>
+		<a href="/tag_names/edit/<?php echo $tn['TagName']['id']; ?>">Edit</a> 
+		<?php if($_SESSION['loggedInUser']['User']['id']==72){ ?>
+			| 
+			<a href="/tag_names/delete/<?php echo $tn['TagName']['id']; ?>">Delete</a>
+		<?php } ?>
 	<?php } ?>
 				</div>
         <div class="existing-tags-list">
